requirements.txt
langchain==0.3.25
langgraph==0.2.70
langchain-google-genai==2.0.10
langchain-ollama==0.2.3
langchain-core==0.3.58
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
PyGithub==2.6.1
pydantic==2.11.3
pyyaml==6.0.2
pytest==8.3.5
httpx==0.28.1
